+++ scripts/deploy-verif.ts
